pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-tavily-search"
version = "0.2.0"
description = "MCP Tavily Search Agent"
authors = [{name = "Arben Ademi", email = "arben.ademi@tuta.io"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.2.0",
    "httpx",
    "python-dotenv",
    "tavily-python",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]

[project.urls]
Homepage = "https://github.com/your-username/mcp-tavily-search"

[project.scripts]
mcp-tavily-search = "mcp_tavily_search.server:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--cov=mcp_tavily_search --cov-report=term-missing"
